Organise classes into packages
