+++ src/main/java/leo/Command.java
@@ -1,3 +1,5 @@
+package leo;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) ;
