+++ src/main/java/leo/DeleteCommand.java
@@ -1,3 +1,5 @@
+package leo;
+
 public class DeleteCommand extends Command {
     private int index;
