+++ src/main/java/leo/Event.java
@@ -1,3 +1,5 @@
+package leo;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
