+++ src/main/java/leo/IndexOutOfBounds.java
@@ -1,3 +1,5 @@
+package leo;
+
 public class IndexOutOfBounds extends Exception {
     public IndexOutOfBounds(String msg) {
         super(msg);
