+++ src/main/java/leo/NoTask.java
@@ -1,3 +1,5 @@
+package leo;
+
 public class NoTask extends Exception {
     public NoTask(String message) {super(message);}
 }
