+++ src/main/java/leo/Parser.java
@@ -1,3 +1,5 @@
+package leo;
+
 public class Parser {
     public Parser() {
