+++ src/main/java/leo/Storage.java
@@ -1,11 +1,11 @@
+package leo;
+
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.nio.file.Files;
-import java.util.List;
 
 
 public class Storage {
