+++ src/main/java/leo/Task.java
@@ -1,3 +1,5 @@
+package leo;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
