+++ src/main/java/leo/UnknownCommand.java
@@ -1,3 +1,5 @@
+package leo;
+
 public class UnknownCommand extends Exception{
     public UnknownCommand(String msg) {super(msg);}
 }
