+++ src/main/java/leo/UnmarkCommand.java
@@ -1,19 +1,21 @@
-public class MarkCommand extends Command {
+package leo;
+
+public class UnmarkCommand extends Command {
     private int index;
 
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
-    // need an instance of an object to run this
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
-            tasks.markDone(index);
-            ui.showMarked(tasks.elem(index));
+            tasks.markUndone(index);
+            ui.showUnmarked(tasks.elem(index));
             storage.save(tasks);
         } catch (Exception err) {
             ui.showError(err);
         }
     }
 }
+
