+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ----------------------------------------
-Hello I'm Leo
+Hello I'm leo.Leo
 What can I do for you?
 ----------------------------------------
 ----------------------------------------
